test-requirements.txt
hacking
stestr
mock
requests-mock

# Docs
sphinx
